Change back to ints

+++ backend/models/office_hours/ticket_statistics.py
@@ -19,8 +19,8 @@ class OfficeHoursTicketStatistics(BaseModel):
     of the `OfficeHoursTicket` database in the PostgreSQL database.
     """
 
-    average_wait_time: float
-    average_duration: float
+    average_wait_time: int
+    average_duration: int
     total_conceptual: int
     total_assignment: int
     total_tickets: int
